ote_velum_classification_final/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
numpy>=1.24.0
pillow>=9.5.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
requests>=2.31.0

# Optional: For model upload/download
# google-api-python-client>=2.0.0
# google-auth-httplib2>=0.1.0
# google-auth-oauthlib>=0.5.0
# huggingface-hub>=0.16.0
